runtime: Merge master revision 19185.
This revision renames several files in the runtime directory
from .c to .goc.


git-svn-id: svn+ssh://gcc.gnu.org/svn/gcc/trunk@212472 138bc75d-0d04-0410-961f-82ee72b054a4

+++ libgo/runtime/print.c
@@ -320,7 +320,7 @@ runtime_printhex(uint64 v)
 void
 runtime_printpointer(void *p)
 {
-	runtime_printhex((uint64)(uintptr)p);
+	runtime_printhex((uintptr)p);
 }
 
 void
